+++ ietfdata/datatracker.py
@@ -1656,7 +1656,7 @@ class DataTracker:
             cache_dir      -- If set, use this directory as a cache for Datatracker objects
         """
         self.session  = requests.Session()
-        self.ua       = "glasgow-ietfdata/0.2.0"          # Update when making a new relaase
+        self.ua       = "glasgow-ietfdata/0.3.0"          # Update when making a new relaase
         self.base_url = "https://datatracker.ietf.org"
         self.http_req = 0
         self.cache_dir = cache_dir