+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log -- ietfdata
 ## v0.2.1 -- ??
 
  - Add API coverage for reviews and IPR disclosures
+ - Update Submission class to use datetime objects where appropriate
 
 
 ## v0.2.0 -- 2020-05-17